tipos BigDecimal e LocalDate alterados para Double e Date

+++ src/main/java/bolicho/model/Produto.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +11,7 @@ public class Produto {
 
     private int id;
     private String descricao;
-    private BigDecimal precoUnitario;
+    private double precoUnitario;
     private String unidadeMedida;
     private double qtdEstoque;
     private boolean arquivado;
